ddraw: Expose the RGB device for version 1.
Based on a patch by Andrew D'Addesio.

The comment dates back to a872c781ea9908bd549b59629585e4fd23ef2211, but at that
time it referred to the reference device. This was changed to the RGB device by
d9fef10b71b9add27cfaeffa6d7e3ab42342bba7.

The reference device is not enumerated by default even on Windows 98, as the
neighbouring comment explains. It would make sense if the affected games do not
function with the reference device, but are fine with the RGB device. It's not
clear whether this is the case, though. According to Andrew D'Addesio in bug
19153, Motoracer 2 does not break even if the reference device is enumerated
(so possibly the bug has to do with some other part of the caps later changed).
As for AvP, I was unable to locate a copy of the game to test.