WIP: include/winuser.rh: flag for demanding X11 window
Adding a WS_* flag that's telling WinE to always create a native (eg. X11)
window on CreateWindowEx*(). Using a reserved bit within the window's style
field, so it's only meaningful when running under WinE. By win32 API spec,
the lower bits of the style field are explicitly reserved and ignored
(probably even cleared out). WineE also clear them out later, but all we need
is the flag reaching CreateWindowEx*(), so we can act on it here.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>