dxgi: Add some D3DKMT forwarded exports.
Those aren't in the Win10 DLL anymore, but they were in the Win7 and
Win8 DLLs. The exports are also in the MinGW runtime, and using a D3DKMT
function in a program built with MinGW might pull it from dxgi, so
having the exports in the Wine DLL is useful.